refactor(desktop): extract DesktopRuntime, slim main.rs to uc_tauri::run
Split the desktop host responsibilities so uc-desktop holds the
GUI-framework agnostic core and uc-tauri owns only the Tauri-specific
last mile.

Moved into uc-desktop (zero tauri dependency):
- DesktopRuntime: facade + task_registry + settings + storage +
  event_emitter + device_id (was uc_tauri::AppRuntime)
- DaemonBootstrapOwnership state for sidecar spawn coordination
- Background task scheduling: file_cache_cleanup + gui_pairing_lease
  (replaced tauri::async_runtime::spawn with tokio::spawn)

uc-tauri now wraps with TauriAppRuntime which adds the Tauri AppHandle
slot, and the bin's run loop lives in uc_tauri::run. Plugins
(mac_rounded_corners) moved from src-tauri/src/plugins to
uc-tauri/src/plugins. main.rs went from 451 lines to 17.

Verified `cargo tree -p uc-desktop` still shows zero tauri / tauri-plugin
edges, satisfying the GUI-agnostic hard constraint.
