+++ src-tauri/crates/uc-tauri/src/commands/mod.rs
@@ -20,7 +20,7 @@ pub fn get_tauri_pid() -> u32 {
 /// Get the stable local device identifier used for telemetry correlation.
 #[tauri::command]
 pub async fn get_device_id(
-    runtime: tauri::State<'_, std::sync::Arc<crate::bootstrap::AppRuntime>>,
+    runtime: tauri::State<'_, std::sync::Arc<crate::bootstrap::TauriAppRuntime>>,
     _trace: Option<TraceMetadata>,
 ) -> Result<String, CommandError> {
     Ok(runtime.device_id())
